Move local development to Docker Compose for prod parity
Local dev previously used `make run-dry` which ran the Go binary directly
on the host, skipping the Chrome sidecar, MCP servers, and full container
environment. This caused bugs with prompt paths, orphaned sessions, and
handoff parsing.

- Add .env.example and docker-compose.override.yaml.example templates
- Add root .gitignore for build artifacts and dev-local state
- Clean up docker-compose.yaml (move dev comments to override template)
- Pass ANTHROPIC_BASE_URL through for LiteLLM/proxy support
- Fix Dockerfile: golang 1.23 -> 1.24, copy go.sum for layer caching
- Replace make run/run-dry with make dev/dev-up/dev-down/dev-logs/dev-rebuild
- Fix --output-format=stream-json requiring --verbose (now unconditional)
- Remove dead Verbose config field
- Fix prompt file read error not finalizing session
- Fix CheckResult.ResponseTimeMs type (int -> float64)
- Update README.md and CLAUDE.md for new workflow

+++ internal/session/handoff.go
@@ -27,7 +27,7 @@ type CheckResult struct {
 	CheckType      string `json:"check_type"`
 	Status         string `json:"status"`
 	Error          string `json:"error,omitempty"`
-	ResponseTimeMs int    `json:"response_time_ms,omitempty"`
+	ResponseTimeMs float64 `json:"response_time_ms,omitempty"`
 }
 
 // handoffFileName is the well-known file name for the handoff payload.
